Move exception to sdk_http
